requirements.txt
python-docx>=1.1.0
google-generativeai
python-dotenv
openai
streamlit
